refactor: move process cache from rest to service layer
