+++ service/src/main/java/io/camunda/service/cache/ProcessCache.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.rest.cache;
+package io.camunda.service.cache;
 
 import com.github.benmanes.caffeine.cache.CacheLoader;
 import com.github.benmanes.caffeine.cache.Caffeine;
@@ -14,8 +14,6 @@ import io.camunda.search.entities.FlowNodeInstanceEntity;
 import io.camunda.search.entities.UserTaskEntity;
 import io.camunda.zeebe.broker.client.api.BrokerTopologyListener;
 import io.camunda.zeebe.broker.client.api.BrokerTopologyManager;
-import io.camunda.zeebe.gateway.rest.config.GatewayRestConfiguration;
-import io.camunda.zeebe.gateway.rest.util.ProcessElementProvider;
 import io.camunda.zeebe.util.cache.CaffeineCacheStatsCounter;
 import io.micrometer.core.instrument.MeterRegistry;
 import java.util.Collections;
@@ -34,9 +32,6 @@ import org.slf4j.LoggerFactory;
  *
  * <p>Use the {@link ProcessCache#getCacheItem(long)} method to load one item or the {@link
  * ProcessCache#getCacheItems(Set)} method to load multiple cache items at once.
- *
- * <p>The process cache default configuration can be changed via the {@link
- * GatewayRestConfiguration.ProcessCacheConfiguration} properties.
  */
 public class ProcessCache {
 
@@ -46,17 +41,16 @@ public class ProcessCache {
   private final ProcessElementProvider processElementProvider;
 
   public ProcessCache(
-      final GatewayRestConfiguration configuration,
+      final Configuration configuration,
       final ProcessElementProvider processElementProvider,
       final BrokerTopologyManager brokerTopologyManager,
       final MeterRegistry meterRegistry) {
     this.processElementProvider = processElementProvider;
+
     final var statsCounter = new CaffeineCacheStatsCounter(NAMESPACE, "process", meterRegistry);
     final var cacheBuilder =
-        Caffeine.newBuilder()
-            .maximumSize(configuration.getProcessCache().getMaxSize())
-            .recordStats(() -> statsCounter);
-    final var expirationIdle = configuration.getProcessCache().getExpirationIdleMillis();
+        Caffeine.newBuilder().maximumSize(configuration.maxSize()).recordStats(() -> statsCounter);
+    final var expirationIdle = configuration.expirationIdleMillis();
     if (expirationIdle != null && expirationIdle > 0) {
       cacheBuilder.expireAfterAccess(expirationIdle, TimeUnit.MILLISECONDS);
     }
@@ -95,6 +89,16 @@ public class ProcessCache {
 
   public void invalidate() {
     cache.invalidateAll();
+  }
+
+  public LoadingCache<Long, ProcessCacheItem> getRawCache() {
+    return cache;
+  }
+
+  public record Configuration(long maxSize, Long expirationIdleMillis) {
+    static Configuration getDefault() {
+      return new Configuration(100, null);
+    }
   }
 
   private final class ProcessCacheLoader implements CacheLoader<Long, ProcessCacheItem> {
