+++ service/src/main/java/io/camunda/service/cache/ProcessCacheItem.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.rest.cache;
+package io.camunda.service.cache;
 
 import java.util.Collections;
 import java.util.Map;
