+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/ElementInstanceController.java
@@ -14,6 +14,7 @@ import io.camunda.search.query.FlowNodeInstanceQuery;
 import io.camunda.security.auth.CamundaAuthenticationProvider;
 import io.camunda.service.ElementInstanceServices;
 import io.camunda.service.ElementInstanceServices.SetVariablesRequest;
+import io.camunda.service.cache.ProcessCache;
 import io.camunda.zeebe.gateway.protocol.rest.ElementInstanceResult;
 import io.camunda.zeebe.gateway.protocol.rest.ElementInstanceSearchQuery;
 import io.camunda.zeebe.gateway.protocol.rest.ElementInstanceSearchQueryResult;
@@ -25,7 +26,6 @@ import io.camunda.zeebe.gateway.rest.SearchQueryResponseMapper;
 import io.camunda.zeebe.gateway.rest.annotation.CamundaGetMapping;
 import io.camunda.zeebe.gateway.rest.annotation.CamundaPostMapping;
 import io.camunda.zeebe.gateway.rest.annotation.CamundaPutMapping;
-import io.camunda.zeebe.gateway.rest.cache.ProcessCache;
 import java.util.concurrent.CompletableFuture;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
