+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/ElementInstanceQueryControllerTest.java
@@ -26,10 +26,10 @@ import io.camunda.search.sort.FlowNodeInstanceSort;
 import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.security.auth.CamundaAuthenticationProvider;
 import io.camunda.service.ElementInstanceServices;
+import io.camunda.service.cache.ProcessCache;
+import io.camunda.service.cache.ProcessCacheItem;
 import io.camunda.zeebe.gateway.protocol.rest.ElementInstanceStateEnum;
 import io.camunda.zeebe.gateway.rest.RestControllerTest;
-import io.camunda.zeebe.gateway.rest.cache.ProcessCache;
-import io.camunda.zeebe.gateway.rest.cache.ProcessCacheItem;
 import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.List;
