+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/UserTaskQueryControllerTest.java
@@ -29,9 +29,9 @@ import io.camunda.search.sort.UserTaskSort;
 import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.security.auth.CamundaAuthenticationProvider;
 import io.camunda.service.UserTaskServices;
+import io.camunda.service.cache.ProcessCache;
+import io.camunda.service.cache.ProcessCacheItem;
 import io.camunda.zeebe.gateway.rest.RestControllerTest;
-import io.camunda.zeebe.gateway.rest.cache.ProcessCache;
-import io.camunda.zeebe.gateway.rest.cache.ProcessCacheItem;
 import java.io.IOException;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
